ADDED: a transformation function to the resource consumption action ALTERED: GreyfishMathExpression and companions moved to a separate package 'core.eval' ADDED: GreyfishMathExpression accepts argument which will be added to the variable resolver

+++ src/main/java/org/asoem/greyfish/core/individual/Agent.java
@@ -294,7 +294,7 @@ public class Agent extends GFAgentDecorator implements IndividualInterface, Movi
     }
 
     @Override
-    public float getAge() {
+    public int getAge() {
         assert simulation.getSteps() >= getTimeOfBirth();
         return simulation.getSteps() - getTimeOfBirth();
     }
